メイン処理ファイルからテスト実行用コードを削除 game.rbのテストコードを分離
